Implemented handling errors for report endpoints

+++ src/report.py
@@ -4,6 +4,7 @@ from src.database import Reports
 from src.generation import generate_xslt_evaluation, generate_schema_evaluation, generate_wellformedness_evaluation, generate_diagnostic_list
 from peewee import DoesNotExist
 from src.constants import ADMIN_TOKEN, PEPPOL_EXECUTABLE, SYNTAX_EXECUTABLE
+from src.error import *
 
 
 def report_wellformedness_v1(invoice_text: str) -> Evaluation:
@@ -12,16 +13,25 @@ def report_wellformedness_v1(invoice_text: str) -> Evaluation:
     return Evaluation(**evaluation.to_json())
 
 def report_schema_v1(invoice_text: str) -> Evaluation:
+    if not report_wellformedness_v1(invoice_text).is_valid:
+        raise InputError(status_code=400, detail="Invoice is not wellformed")
+    
     evaluation = generate_schema_evaluation(invoice_text)
     
     return Evaluation(**evaluation.to_json())
 
 def report_syntax_v1(invoice_text: str) -> Evaluation:
+    if not report_wellformedness_v1(invoice_text).is_valid:
+        raise InputError(status_code=400, detail="Invoice is not wellformed")
+    
     evaluation = generate_xslt_evaluation(SYNTAX_EXECUTABLE, invoice_text)
     
     return Evaluation(**evaluation.to_json())
 
 def report_peppol_v1(invoice_text: str) -> Evaluation:
+    if not report_wellformedness_v1(invoice_text).is_valid:
+        raise InputError(status_code=400, detail="Invoice is not wellformed")
+    
     evaluation = generate_xslt_evaluation(PEPPOL_EXECUTABLE, invoice_text)
 
     return Evaluation(**evaluation.to_json())
@@ -38,13 +48,19 @@ def report_list_by_v1(order_by: OrderBy) -> List[int]:
     return ReportIDs(report_ids=[report.id for report in Reports.select().order_by(order)])
 
 def report_change_name_v1(token: str, report_id: int, new_name: str) -> Dict[None, None]:
+    if len(new_name) > 100:
+        raise InputError(status_code=400, detail="New name is longer than 100 characters")
+    
     if not token == ADMIN_TOKEN:
-        raise Exception("Only admins can change the names of reports at the moment")
+        raise InputError(status_code=400, detail="Only admins can change the names of reports at the moment")
+    
+    if report_id < 0:
+        raise InputError(status_code=400, detail="Report id cannot be less than 0")
     
     try:
         report = Reports.get_by_id(report_id)
     except DoesNotExist:
-        raise Exception(f"Report with id {report_id} not found")
+        raise NotFoundError(status_code=404, detail=f"Report with id {report_id} not found")
     
     report.invoice_name = new_name
     report.save()
@@ -52,13 +68,16 @@ def report_change_name_v1(token: str, report_id: int, new_name: str) -> Dict[Non
     return {}
 
 def report_delete_v1(token: str, report_id: int) -> Dict[None, None]:
+    if report_id < 0:
+        raise InputError(status_code=400, detail="Report id cannot be less than 0")
+    
     if not token == ADMIN_TOKEN:
-        raise Exception("Only admins can delete reports at the moment")
+        raise InputError(status_code=400, detail="Only admins can change the names of reports at the moment")
     
     try:
         report = Reports.get_by_id(report_id)
     except DoesNotExist:
-        raise Exception(f"Report with id {report_id} not found")
+        raise NotFoundError(status_code=404, detail=f"Report with id {report_id} not found")
     
     report.delete_instance()
